UniProtParser::Extractor: no longer parse CC lines
We no longer use comments for anything in the parser so don't waste
memory and CPU cycles on parsing them.

Note that this now causes perlcritic to complain about the private
method _get_comments() being unused. Either write an unit test which
makes sure this method doesn't succumb to bit rot or remove it.