UniProtParser::Extractor: allow curly brackets in descriptions
Turns out there *are* protein names containing curly brackets, e.g.
(4-{4-[2-(gamma-L-glutamylamino)ethyl]phenoxymethyl}furan-2-yl)methanamine synthase .
There aren't many of those, for instance there are only 2 in the current Swiss-Prot
file - but they are there. Relax the regular-expression pattern to allow this.

One upshot of this has been that now the stripping of the evidence-code block
is handled by a separate substitution regex. This might perhaps be a bit less
efficient, then again this way is more readable and that part of the match was
fragile to begin with (it depended on a space before the opening curly bracket).